refact!: Merge `api/request.js` into `Api` class
